Project restructuring
- Restructured the project a bit to move the code into a structured folder setup
